+++ satellite/src/main/java/com/linbit/linstor/core/objects/EbsRemoteSatelliteFactory.java
@@ -94,7 +94,6 @@ public class EbsRemoteSatelliteFactory
             if (!remote.getUuid().equals(uuid))
             {
                 CriticalError.dieUuidMissmatch(
-                    errorReporter,
                     S3Remote.class.getSimpleName(),
                     remote.getName().displayValue,
                     remoteNameRef.displayValue,