+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+ - Fixed missing transaction-commits during evacuate's copy-snap
+
 ## [1.33.0] - 2025-12-11
 
 ### Changed