+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed missing transaction-commits during evacuate's copy-snap
+- Make sure updateDiscGranularity is not called if only DRBD_DELETE is set
 - Also delete error-report text files on error delete api call
 
 ## [1.33.0] - 2025-12-11