+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.33.1] - 2025-12-22
+
 ### Fixed
 
 - Fixed missing transaction-commits during evacuate's copy-snap